# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerLens API",
    "version": "1.0.0",
    "description": "Session-authenticated HTTP API for the CareerLens web application."
  },
  "paths": {
...
    "/api/v1/feed": {
      "get": {
        "operationId": "config_api_v1_feed_feed",
        "summary": "List vacancies",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor returned by the previous response.",
              "title": "Cursor"
            },
            "required": false,
            "description": "Opaque cursor returned by the previous response."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 20,
              "description": "Number of vacancies to return.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false,
            "description": "Number of vacancies to return."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedOut"
                },
                "example": {
                  "items": [
                    {
                      "id": 42,
                      "title": "Python Developer",
                      "company": "Acme",
                      "location": "Remote",
                      "posted_date": "2026-08-21",
                      "scraped_at": "2026-08-21T10:30:00Z",
                      "source_updated_at": "2026-08-21T11:45:00Z",
                      "is_deftech": true,
                      "source": {
                        "code": "dou",
                        "name": "DOU",
                        "icon_url": "/source-icons/dou.png"
                      },
                      "url": "https://example.com/jobs/42",
                      "saved": true,
                      "hidden": false,
                      "seen": true,
                      "match": {
                        "score": 86,
                        "reason": "Strong Python and Django overlap.",
                        "evidence": {
                          "items": [
                            {
                              "type": "strong",
                              "label": "Python",
                              "explanation": "Five years of experience."
                            }
                          ],
                          "evidence_coverage": 0.86
                        },
                        "precise": true,
                        "scored_at": "2026-08-21T12:00:00Z"
                      }
                    }
                  ],
                  "next_cursor": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                },
                "example": {
                  "detail": "Unauthorized"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "description": "Returns vacancies in stable newest-first order.",
        "tags": [
          "feed"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/v1/feed/{vacancy_id}": {
      "get": {
        "operationId": "config_api_v1_feed_feed_detail",
        "summary": "Get vacancy detail",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "schema": {
              "description": "CareerLens vacancy identifier.",
              "title": "Vacancy Id",
              "type": "integer"
            },
            "required": true,
            "description": "CareerLens vacancy identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyDetailOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "description": "Returns one vacancy with its source description and the current user's match.",
        "tags": [
          "feed"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/v1/feed/{vacancy_id}/saved": {
      "post": {
        "operationId": "config_api_v1_feed_set_saved",
        "summary": "Set saved state",
        "parameters": [
          {
            "in": "cookie",
            "name": "csrftoken",
            "schema": {
              "description": "CSRF cookie issued by GET /api/v1/me.",
              "title": "Csrftoken",
              "type": "string"
            },
            "required": true,
            "description": "CSRF cookie issued by GET /api/v1/me."
          },
          {
            "in": "header",
            "name": "X-CSRFToken",
            "schema": {
              "description": "Value matching the csrftoken cookie.",
              "title": "X-Csrftoken",
              "type": "string"
            },
            "required": true,
            "description": "Value matching the csrftoken cookie."
          },
          {
            "in": "path",
            "name": "vacancy_id",
            "schema": {
              "description": "CareerLens vacancy identifier.",
              "title": "Vacancy Id",
              "type": "integer"
            },
            "required": true,
            "description": "CareerLens vacancy identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "description": "Sets the current user's saved flag idempotently. Requires a valid CSRF token.",
        "tags": [
          "feed"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedIn"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/v1/feed/{vacancy_id}/hidden": {
      "post": {
        "operationId": "config_api_v1_feed_set_hidden",
        "summary": "Set hidden state",
        "parameters": [
          {
            "in": "cookie",
            "name": "csrftoken",
            "schema": {
              "description": "CSRF cookie issued by GET /api/v1/me.",
              "title": "Csrftoken",
              "type": "string"
            },
            "required": true,
            "description": "CSRF cookie issued by GET /api/v1/me."
          },
          {
            "in": "header",
            "name": "X-CSRFToken",
            "schema": {
              "description": "Value matching the csrftoken cookie.",
              "title": "X-Csrftoken",
              "type": "string"
            },
            "required": true,
            "description": "Value matching the csrftoken cookie."
          },
          {
            "in": "path",
            "name": "vacancy_id",
            "schema": {
              "description": "CareerLens vacancy identifier.",
              "title": "Vacancy Id",
              "type": "integer"
            },
            "required": true,
            "description": "CareerLens vacancy identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HiddenOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "description": "Sets the current user's hidden flag idempotently. Requires a valid CSRF token.",
        "tags": [
          "feed"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HiddenIn"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    },
    "/api/v1/feed/{vacancy_id}/seen": {
      "post": {
        "operationId": "config_api_v1_feed_mark_seen",
        "summary": "Mark vacancy seen",
        "parameters": [
          {
            "in": "cookie",
            "name": "csrftoken",
            "schema": {
              "description": "CSRF cookie issued by GET /api/v1/me.",
              "title": "Csrftoken",
              "type": "string"
            },
            "required": true,
            "description": "CSRF cookie issued by GET /api/v1/me."
          },
          {
            "in": "header",
            "name": "X-CSRFToken",
            "schema": {
              "description": "Value matching the csrftoken cookie.",
              "title": "X-Csrftoken",
              "type": "string"
            },
            "required": true,
            "description": "Value matching the csrftoken cookie."
          },
          {
            "in": "path",
            "name": "vacancy_id",
            "schema": {
              "description": "CareerLens vacancy identifier.",
              "title": "Vacancy Id",
              "type": "integer"
            },
            "required": true,
            "description": "CareerLens vacancy identifier."
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "description": "Marks the vacancy seen once for the current user. Requires a valid CSRF token.",
        "tags": [
          "feed"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "examples": [
          {
            "detail": "Unauthorized"
          }
        ],
        "properties": {
          "detail": {
            "description": "Human-readable reason the request was rejected.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "FeedOut": {
        "properties": {
          "items": {
            "description": "Vacancies in stable newest-first order.",
            "items": {
              "$ref": "#/components/schemas/VacancyOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor for the next page, or null at the end.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "FeedOut",
        "type": "object"
      },
...
      "VacancyOut": {
        "examples": [
          {
            "company": "Acme",
            "hidden": false,
            "id": 42,
            "is_deftech": true,
            "location": "Remote",
            "match": {
              "evidence": {
                "evidence_coverage": 0.86,
                "items": [
                  {
                    "explanation": "Five years of experience.",
                    "label": "Python",
                    "type": "strong"
                  }
                ]
              },
              "precise": true,
              "reason": "Strong Python and Django overlap.",
              "score": 86,
              "scored_at": "2026-08-21T12:00:00Z"
            },
            "posted_date": "2026-08-21",
            "saved": true,
            "scraped_at": "2026-08-21T10:30:00Z",
            "seen": true,
            "source": {
              "code": "dou",
              "icon_url": "/source-icons/dou.png",
              "name": "DOU"
            },
            "source_updated_at": "2026-08-21T11:45:00Z",
            "title": "Python Developer",
            "url": "https://example.com/jobs/42"
          }
        ],
        "properties": {
          "id": {
            "description": "CareerLens vacancy identifier.",
            "title": "Id",
            "type": "integer"
          },
          "title": {
            "description": "Vacancy title.",
            "title": "Title",
            "type": "string"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Company name, or null when unavailable.",
            "title": "Company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Vacancy location, or null when unavailable.",
            "title": "Location"
          },
          "posted_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source publication date, or null when unavailable.",
            "title": "Posted Date"
          },
          "scraped_at": {
            "description": "Time when CareerLens imported the vacancy.",
            "format": "date-time",
            "title": "Scraped At",
            "type": "string"
          },
          "source_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last source update time, or null when unavailable.",
            "title": "Source Updated At"
          },
          "is_deftech": {
            "description": "Whether the vacancy or its company is classified as DefTech.",
            "title": "Is Deftech",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/SourceOut",
            "description": "Vacancy source."
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Original vacancy URL, or null when unavailable.",
            "title": "Url"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Current user's match result, or null when not scored."
          },
          "saved": {
            "description": "Whether the current user saved the vacancy.",
            "title": "Saved",
            "type": "boolean"
          },
          "hidden": {
            "description": "Whether the current user hid the vacancy.",
            "title": "Hidden",
            "type": "boolean"
          },
          "seen": {
            "description": "Whether the current user opened the vacancy.",
            "title": "Seen",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "location",
          "posted_date",
          "scraped_at",
          "source_updated_at",
          "is_deftech",
          "source",
          "url",
          "match",
          "saved",
          "hidden",
          "seen"
        ],
        "title": "VacancyOut",
        "type": "object"
      },
      "VacancyDetailOut": {
        "examples": [
          {
            "company": "Acme",
            "hidden": false,
            "id": 42,
            "is_deftech": true,
            "location": "Remote",
            "match": {
              "evidence": {
                "evidence_coverage": 0.86,
                "items": [
                  {
                    "explanation": "Five years of experience.",
                    "label": "Python",
                    "type": "strong"
                  }
                ]
              },
              "precise": true,
              "reason": "Strong Python and Django overlap.",
              "score": 86,
              "scored_at": "2026-08-21T12:00:00Z"
            },
            "posted_date": "2026-08-21",
            "saved": true,
            "scraped_at": "2026-08-21T10:30:00Z",
            "seen": true,
            "source": {
              "code": "dou",
              "icon_url": "/source-icons/dou.png",
              "name": "DOU"
            },
            "source_updated_at": "2026-08-21T11:45:00Z",
            "title": "Python Developer",
            "url": "https://example.com/jobs/42"
          }
        ],
        "properties": {
          "id": {
            "description": "CareerLens vacancy identifier.",
            "title": "Id",
            "type": "integer"
          },
          "title": {
            "description": "Vacancy title.",
            "title": "Title",
            "type": "string"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Company name, or null when unavailable.",
            "title": "Company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Vacancy location, or null when unavailable.",
            "title": "Location"
          },
          "posted_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source publication date, or null when unavailable.",
            "title": "Posted Date"
          },
          "scraped_at": {
            "description": "Time when CareerLens imported the vacancy.",
            "format": "date-time",
            "title": "Scraped At",
            "type": "string"
          },
          "source_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last source update time, or null when unavailable.",
            "title": "Source Updated At"
          },
          "is_deftech": {
            "description": "Whether the vacancy or its company is classified as DefTech.",
            "title": "Is Deftech",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/SourceOut",
            "description": "Vacancy source."
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Original vacancy URL, or null when unavailable.",
            "title": "Url"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Current user's match result, or null when not scored."
          },
          "saved": {
            "description": "Whether the current user saved the vacancy.",
            "title": "Saved",
            "type": "boolean"
          },
          "hidden": {
            "description": "Whether the current user hid the vacancy.",
            "title": "Hidden",
            "type": "boolean"
          },
          "seen": {
            "description": "Whether the current user opened the vacancy.",
            "title": "Seen",
            "type": "boolean"
          },
          "description": {
            "description": "Vacancy description safe for direct text rendering.",
            "title": "Description",
            "type": "string"
          },
          "description_status": {
            "description": "Description representation.",
            "enum": [
              "markdown",
              "source"
            ],
            "title": "Description Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "location",
          "posted_date",
          "scraped_at",
          "source_updated_at",
          "is_deftech",
          "source",
          "url",
          "match",
          "saved",
          "hidden",
          "seen",
          "description",
          "description_status"
        ],
        "title": "VacancyDetailOut",
        "type": "object"
      },
      "SavedOut": {
        "examples": [
          {
            "saved": true
          }
        ],
        "properties": {
          "saved": {
            "description": "Confirmed saved state.",
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "saved"
        ],
        "title": "SavedOut",
        "type": "object"
      },
      "SavedIn": {
        "examples": [
          {
            "saved": true
          }
        ],
        "properties": {
          "saved": {
            "description": "Whether the vacancy is saved.",
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "saved"
        ],
        "title": "SavedIn",
        "type": "object"
      },
      "HiddenOut": {
        "examples": [
          {
            "hidden": true
          }
        ],
        "properties": {
          "hidden": {
            "description": "Confirmed hidden state.",
            "title": "Hidden",
            "type": "boolean"
          }
        },
        "required": [
          "hidden"
        ],
        "title": "HiddenOut",
        "type": "object"
      },
      "HiddenIn": {
        "examples": [
          {
            "hidden": true
          }
        ],
        "properties": {
          "hidden": {
            "description": "Whether the vacancy is hidden.",
            "title": "Hidden",
            "type": "boolean"
          }
        },
        "required": [
          "hidden"
        ],
        "title": "HiddenIn",
        "type": "object"
      }
    },
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  },
  "servers": [],
...
}